pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = ""
authors = [
    {name = "damodara",email = "jay108@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=8.0.0,<9.0.0)",
    "djangorestframework-stubs (>=3.16.8,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
